// src/types/apis/candle.ts
export type CandleMinute = {
 market: string
 candle_date_time_utc: string
 candle_date_time_kst: string
 opening_price: number
 high_price: number
 low_price: number
 trade_price: number
 timestamp: number
 candle_acc_trade_price: number
 candle_acc_trade_volume: number
 unit: number
}

export type DailyCandle = {
 market: string
 candle_date_time_utc: string
 candle_date_time_kst: string
 opening_price: number
 high_price: number
 low_price: number
 trade_price: number
 timestamp: number
 candle_acc_trade_price: number
 candle_acc_trade_volume: number
 prev_closing_price: number
 change_price: number
 change_rate: number
}